feat: move tests domains and pojo to v1 subpackage
