+++ src/test/java/it/pagopa/ecommerce/commons/domain/v1/EmailTest.java
@@ -1,6 +1,5 @@
-package it.pagopa.ecommerce.commons.domain;
+package it.pagopa.ecommerce.commons.domain.v1;
 
-import it.pagopa.ecommerce.commons.domain.v1.Email;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
